Build script changes
- Make script faster by removing sleep.
- Make script smaller by using a single function to build both targets.
- Make script smarter by making BASE_DIR path relative to script path. Now you can checkout to any directory.
- Make script portable by removing terminal stuff such as text colors when running from a non terminal app such as BBEdit.app.
- Make script output cleaner by removing clear - which had the effect of adding a bunch of blank lines to the terminal or an error in BBEdit.app.
- Now you can choose which targets to build - building only one target is much faster than building both.
- Instead of copying RefindPlusPkg to edk2, we'll just make a symbolic link.
- Remove checkout. No need to do that before every build.

+++ 000-BuildScript/RefindPlusBuilder.sh
@@ -8,18 +8,47 @@
  # MIT License
 ###
 
+# Parse parameters, setup colors if terminal
+BUILD_REL=1
+BUILD_DBG=1
+if [[ $1 == -dbg ]]; then
+    BUILD_REL=0
+    shift
+fi
+if [[ $1 == -rel ]]; then
+    BUILD_DBG=0
+    shift
+fi
+
+color_base=""
+color_info=""
+color_status=""
+color_error=""
+color_normal=""
+
+if test -t 1; then
+    ncolors=$(tput colors)
+    if test -n "$ncolors" && test $ncolors -ge 8; then
+        color_base="\033[0;36m"
+        color_info="\033[0;33m"
+        color_status="\033[0;32m"
+        color_error="\033[0;31m"
+        color_normal="\033[0m"
+    fi
+fi
+
 # Provide custom colours
 msg_base() {
-    echo -e "\033[0;36m$1\033[0m"
+    echo -e "$color_base$1$color_normal"
 }
 msg_info() {
-    echo -e "\033[0;33m$1\033[0m"
+    echo -e "$color_info$1$color_normal"
 }
 msg_status() {
-    echo -e "\033[0;32m$1\033[0m"
+    echo -e "$color_status$1$color_normal"
 }
 msg_error() {
-    echo -e "\033[0;31m$1\033[0m"
+    echo -e "$color_error$1$color_normal"
 }
 
 ## ERROR HANDLER ##
@@ -38,12 +67,9 @@ trap runErr ERR
 
 
 # Set things up for build
-clear
 msg_info '## RefindPlusBuilder - Setting Up ##'
 msg_info '------------------------------------'
-sleep 2
-EDIT_BRANCH="${1:-GOPFix}"
-BASE_DIR="${HOME}/Documents/RefindPlus"
+BASE_DIR="$(cd "${BASH_SOURCE%/*}/../../"; pwd)"
 WORK_DIR="${BASE_DIR}/Working"
 EDK2_DIR="${BASE_DIR}/edk2"
 if [ ! -d "${EDK2_DIR}" ] ; then
@@ -62,55 +88,36 @@ GLOBAL_FILE_TMP_DBG="${EDK2_DIR}/RefindPlusPkg/MainLoader/globalExtra-DBG.txt"
 BUILD_DSC="${EDK2_DIR}/RefindPlusPkg/RefindPlusPkg.dsc"
 BUILD_DSC_REL="${EDK2_DIR}/RefindPlusPkg/RefindPlusPkg-REL.dsc"
 BUILD_DSC_DBG="${EDK2_DIR}/RefindPlusPkg/RefindPlusPkg-DBG.dsc"
-OUR_RAND=$(( RANDOM % 7 ))
 BASETOOLS='true'
-if [ -d "${EDK2_DIR}/BaseTools/Source/C/bin" ] && [ "${OUR_RAND}" != "0" ] ; then
+if [ -d "${EDK2_DIR}/BaseTools/Source/C/bin" ] ; then
     BASETOOLS='false'
 fi
 
-pushd "${WORK_DIR}" > /dev/null || exit 1
-msg_base "Checkout '${EDIT_BRANCH}' branch..."
-git checkout ${EDIT_BRANCH} > /dev/null
-msg_status '...OK'; echo ''
-sleep 2
 msg_base 'Update RefindPlusPkg...'
-
-# Remove later #
-rm -fr "${EDK2_DIR}/RefindPkg"
-# Remove later #
-
-rm -fr "${EDK2_DIR}/RefindPlusPkg"
-cp -fa "${WORK_DIR}" "${EDK2_DIR}/RefindPlusPkg"
-rm -fr "${EDK2_DIR}/RefindPlusPkg/.gitignore"
-rm -fr "${EDK2_DIR}/RefindPlusPkg/.git"
+if [ ! -L "${EDK2_DIR}/RefindPlusPkg" ]; then
+	rm -fr "${EDK2_DIR}/RefindPlusPkg"
+    ln -s "${WORK_DIR}" "${EDK2_DIR}/RefindPlusPkg"
+fi
 msg_status '...OK'; echo ''
-sleep 2
-popd > /dev/null || exit 1
 
 if [ "${BASETOOLS}" == 'true' ] ; then
     pushd "${EDK2_DIR}/BaseTools/Source/C" > /dev/null || exit 1
     msg_base 'Make Clean...'
-    sleep 2
     make clean
     msg_status '...OK'; echo ''
     popd > /dev/null || exit 1
 
     pushd "${EDK2_DIR}" > /dev/null || exit 1
-    sleep 2
     msg_base 'Make BaseTools...'
-    sleep 2
     make -C BaseTools/Source/C
     msg_status '...OK'; echo ''
-    sleep 2
     popd > /dev/null || exit 1
 fi
 
 
 # Basic clean up
-clear
 msg_info '## RefindPlusBuilder - Initial Clean Up ##'
 msg_info '------------------------------------------'
-sleep 2
 
 if [ -d "${EDK2_DIR}/Build" ] ; then
     rm -fr "${EDK2_DIR}/Build"
@@ -121,73 +128,57 @@ if [ -d "${OUTPUT_DIR}" ] ; then
 fi
 mkdir -p "${OUTPUT_DIR}"
 
-# Build release version
-clear
-msg_info '## RefindPlusBuilder - Building REL Version ##'
-msg_info '----------------------------------------------'
 
-sleep 2
-pushd "${EDK2_DIR}" > /dev/null || exit 1
-if [ -d "${EDK2_DIR}/.Build-TMP" ] ; then
-    rm -fr "${EDK2_DIR}/.Build-TMP"
-fi
-if [ -f "${GLOBAL_FILE}" ] ; then
-    rm -fr "${GLOBAL_FILE}"
-fi
-cp "${GLOBAL_FILE_TMP_REL}" "${GLOBAL_FILE}"
+DoBuild () {
+    local BuildType="$1"
+    local BuildOption="$2"
+    local GLOBAL_FILE_TMP="$3"
+    local BUILD_DSC_TMP="$4"
+    local BINARY_DIR="$5"
+    
+    #clear
+    msg_info "## RefindPlusBuilder - Building $BuildType Version ##"
+    msg_info '----------------------------------------------'
 
-if [ -f "${BUILD_DSC}" ] ; then
-    rm -fr "${BUILD_DSC}"
-fi
-cp "${BUILD_DSC_REL}" "${BUILD_DSC}"
+    pushd "${EDK2_DIR}" > /dev/null || exit 1
+    if [ -d "${EDK2_DIR}/.Build-TMP" ] ; then
+        rm -fr "${EDK2_DIR}/.Build-TMP"
+    fi
+    if [ -f "${GLOBAL_FILE}" ] ; then
+        rm -fr "${GLOBAL_FILE}"
+    fi
+    cp "${GLOBAL_FILE_TMP}" "${GLOBAL_FILE}"
 
-source edksetup.sh BaseTools
-build -b RELEASE
+    if [ -f "${BUILD_DSC}" ] ; then
+        rm -fr "${BUILD_DSC}"
+    fi
+    cp "${BUILD_DSC_TMP}" "${BUILD_DSC}"
 
-if [ -d "${EDK2_DIR}/Build" ] ; then
-    cp "${BINARY_DIR_REL}/RefindPlus.efi" "${OUTPUT_DIR}/BOOTx64-REL.efi"
-fi
-popd > /dev/null || exit 1
-echo ''
-msg_info "Completed REL Build on '${EDIT_BRANCH}' Branch of RefindPlus"
-sleep 2
-msg_info 'Preparing DBG Build...'
-echo ''
-sleep 4
+    source edksetup.sh BaseTools
+    build -b $BuildOption
+
+    if [ -d "${EDK2_DIR}/Build" ] ; then
+        cp "${BINARY_DIR}/RefindPlus.efi" "${OUTPUT_DIR}/BOOTx64-${BuildType}.efi"
+    fi
+    if [ -d "${EDK2_DIR}/.Build-TMP" ] ; then
+        rm -fr "${EDK2_DIR}/.Build-TMP"
+    fi
+    popd > /dev/null || exit 1
+    echo ''
+    msg_info "Completed ${BuildType} Build on current branch of RefindPlus"
+    echo ''
+}
 
 
-# Build debug version
-clear
-msg_info '## RefindPlusBuilder - Building DBG Version ##'
-msg_info '----------------------------------------------'
-sleep 2
-pushd "${EDK2_DIR}" > /dev/null || exit 1
-if [ -f "${GLOBAL_FILE}" ] ; then
-    rm -fr "${GLOBAL_FILE}"
+if ((BUILD_REL)); then
+    # Build release version
+    DoBuild REL RELEASE "${GLOBAL_FILE_TMP_REL}" "${BUILD_DSC_REL}" "${BINARY_DIR_REL}"
 fi
-cp "${GLOBAL_FILE_TMP_DBG}" "${GLOBAL_FILE}"
 
-if [ -f "${BUILD_DSC}" ] ; then
-    rm -fr "${BUILD_DSC}"
+if ((BUILD_DBG)); then
+    # Build debug version
+    DoBuild DBG DEBUG "${GLOBAL_FILE_TMP_DBG}" "${BUILD_DSC_DBG}" "${BINARY_DIR_DBG}"
 fi
-cp "${BUILD_DSC_DBG}" "${BUILD_DSC}"
-
-source edksetup.sh BaseTools
-build -b DEBUG
-
-if [ -d "${EDK2_DIR}/Build" ] ; then
-    cp "${BINARY_DIR_DBG}/RefindPlus.efi" "${OUTPUT_DIR}/BOOTx64-DBG.efi"
-fi
-if [ -d "${EDK2_DIR}/.Build-TMP" ] ; then
-    rm -fr "${EDK2_DIR}/.Build-TMP"
-fi
-if [ -d "${EDK2_DIR}/Build" ] ; then
-    cp "${BINARY_DIR_REL}/RefindPlus.efi" "${OUTPUT_DIR}/BOOTx64-REL.efi"
-fi
-popd > /dev/null || exit 1
-echo ''
-msg_info "Completed DBG Build on '${EDIT_BRANCH}' Branch of RefindPlus"
-echo ''
 
 
 # Tidy up
@@ -198,7 +189,6 @@ cp "${GLOBAL_FILE_TMP_REL}" "${GLOBAL_FILE}"
 if [ -f "${BUILD_DSC}" ] ; then
     rm -fr "${BUILD_DSC}"
 fi
-cp "${BUILD_DSC_REL}" "${BUILD_DSC}"
 echo ''
 msg_info 'Output EFI Files...'
 msg_status "RefindPlus EFI Files (BOOTx64)      : '${OUTPUT_DIR}'"
